requirements.txt
Flask
openai
gspread
google-auth
speechrecognition
textblob
pyaudio
vosk